Restructure more
