--- docker-compose.yml
@@ -1,21 +1,0 @@
-version: '3'
-
-services:
-    php:
-        container_name: yii-php
-        image: yiisoftware/yii-php:8.0-fpm
-        working_dir: /app
-        volumes:
-            - ./:/app
-    nginx:
-        image: nginx:alpine
-        container_name: yii-nginx
-        ports:
-            - 8080:80
-            - 8081:81
-        volumes:
-            - ./:/app
-            - ./data/nginx/:/etc/nginx/conf.d/
-        depends_on:
-            - php
-        restart: always
