+++ src/Command/HelloCommand.php
@@ -9,7 +9,7 @@ use Symfony\Component\Console\Input\InputInterface;
 use Symfony\Component\Console\Output\OutputInterface;
 use Yiisoft\Yii\Console\ExitCode;
 
-final class Hello extends Command
+final class HelloCommand extends Command
 {
     public const NAME = 'hello';
     protected static $defaultName = self::NAME;
